fix(stats): cast np.corrcoef result to float64 for mypy-strict on py3.10/3.11
np.corrcoef infers floating[Any]; partial_correlation_matrix expects float64, so
strict mypy failed on CI (py3.10/3.11) though it passed on the 3.9 dev box. Cast
the correlation matrix to float64 at the call site. Verified ruff+format+mypy
(strict)+pytest under py3.10. No fabricated data.

+++ src/membench/stats/partial_correlation.py
@@ -199,7 +199,7 @@ def partial_correlation(
     if np.any(matrix.std(axis=0) == 0):
         raise ValueError("every variable must have non-zero variance")
 
-    corr = np.corrcoef(matrix, rowvar=False)
+    corr = np.corrcoef(matrix, rowvar=False).astype(np.float64)
     partial = float(partial_correlation_matrix(corr)[0, 1])
 
     # Semipartial: covariates removed from x only -> corr(y, residual of x on Z).
